# docs/openapi.yaml
openapi: 3.0.0
paths:
  /colleges:
    get:
      operationId: CollegesController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
        - name: search
          required: false
          in: query
          description: Search term for college name or code
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List all colleges with pagination
      tags:
        - Colleges
    post:
      operationId: CollegesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollegeDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Create a new college (Admin only)
      tags:
        - Colleges
  /colleges/{id}:
    get:
      operationId: CollegesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get college by ID
      tags:
        - Colleges
    patch:
      operationId: CollegesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollegeDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update college (Admin only)
      tags:
        - Colleges
    delete:
      operationId: CollegesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete college (Admin only)
      tags:
        - Colleges
  /colleges/{collegeId}/departments:
    get:
      operationId: CollegesController_findDepartments
      parameters:
        - name: collegeId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List departments for a college
      tags:
        - Colleges
  /departments/{id}:
    get:
      operationId: DepartmentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get department by ID
      tags:
        - Departments
    patch:
      operationId: DepartmentsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update department (Admin only)
      tags:
        - Departments
    delete:
      operationId: DepartmentsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete department (Admin only)
      tags:
        - Departments
  /departments:
    post:
      operationId: DepartmentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Create a new department (Admin only)
      tags:
        - Departments
...
  /blogs:
    get:
      operationId: BlogsController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of items to return
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: offset
          required: false
          in: query
          description: Number of items to skip
          schema:
            minimum: 0
            default: 0
            type: number
        - name: search
          required: false
          in: query
          description: Search by blog title
          schema:
            type: string
        - name: tags
          required: false
          in: query
          description: Filter by tags
          schema:
            type: array
            items:
              type: string
        - name: authorId
          required: false
          in: query
          description: Filter by author member ID
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: List published blogs with filters
      tags:
        - Blogs
    post:
      operationId: BlogsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Create a new blog post (Member only)
      tags:
        - Blogs
  /blogs/slug/{slug}:
    get:
      operationId: BlogsController_findBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get published blog by slug
      tags:
        - Blogs
  /blogs/{id}:
    get:
      operationId: BlogsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Get blog by ID (includes unpublished for author/admin)
      tags:
        - Blogs
    patch:
      operationId: BlogsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlogDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Update blog post (Author or Admin only)
      tags:
        - Blogs
    delete:
      operationId: BlogsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Delete blog post (Author or Admin only)
      tags:
        - Blogs
  /blogs/{id}/publish:
    patch:
      operationId: BlogsController_publish
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishBlogDto'
      responses:
        '200':
          description: ''
      security:
        - team-auth: []
      summary: Publish/unpublish blog post (Author or Admin only)
      tags:
        - Blogs
  /certificates/event/{eventId}/participant/{collegeIdNo}:
    get:
      operationId: CertificatesController_getCertificate
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
        - name: collegeIdNo
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      summary: Download certificate PDF for participant (rate limited)
      tags:
        - Certificates
  /certificates/my-certificates:
    get:
      operationId: CertificatesController_getMyCertificates
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - user-auth: []
      summary: Get list of certificates for authenticated participant
      tags:
        - Certificates
  /certificates/bulk-generate:
    post:
      operationId: CertificatesController_bulkGenerate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkGenerateCertificatesDto'
      responses:
        '201':
          description: ''
      security:
        - team-auth: []
      summary: Bulk generate certificates for all event participants (Admin only)
      tags:
        - Certificates
...
info:
  title: The Robotics Club API
  description: API for the college robotics club management platform
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local Development
  - url: https://api.roboticsclub.example.com
    description: Production
components:
  securitySchemes:
    user-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for participants (external users)
    team-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token for club members/admins
  schemas:
    CreateCollegeDto:
      type: object
      properties:
        name:
          type: string
          description: College name
          example: MIT
        code:
          type: string
          description: College code
          example: MIT001
        location:
          type: string
          description: College location
          example: Cambridge, MA
      required:
        - name
        - code
    UpdateCollegeDto:
      type: object
      properties:
        name:
          type: string
          description: College name
        code:
          type: string
          description: College code
        location:
          type: string
          description: College location
    CreateDepartmentDto:
      type: object
      properties:
        name:
          type: string
          description: Department name
          example: Computer Science
        code:
          type: string
          description: Department code
          example: CS
        collegeId:
          type: string
          description: College ID this department belongs to
      required:
        - name
        - code
        - collegeId
    UpdateDepartmentDto:
      type: object
      properties:
        name:
          type: string
          description: Department name
        code:
          type: string
          description: Department code
...
    CreateBlogDto:
      type: object
      properties:
        title:
          type: string
          description: Blog title
          example: Getting Started with Robotics
        slug:
          type: string
          description: URL slug (auto-generated if not provided)
          example: getting-started-robotics
        excerpt:
          type: string
          description: Short excerpt/summary
        content:
          type: object
          description: Rich content (Quill Delta JSON)
        contentHtml:
          type: string
          description: HTML representation of content
        coverImage:
          type: string
          description: Cover image URL
        tags:
          description: Blog tags
          type: array
          items:
            type: string
        published:
          type: boolean
          description: Publish immediately
          default: false
      required:
        - title
        - excerpt
        - content
        - contentHtml
    UpdateBlogDto:
      type: object
      properties:
        title:
          type: string
          description: Blog title
        slug:
          type: string
          description: URL slug
        excerpt:
          type: string
          description: Short excerpt/summary
        content:
          type: object
          description: Rich content (Quill Delta JSON)
        contentHtml:
          type: string
          description: HTML representation of content
        coverImage:
          type: string
          description: Cover image URL
        tags:
          description: Blog tags
          type: array
          items:
            type: string
    PublishBlogDto:
      type: object
      properties:
        published:
          type: boolean
          description: Publish status (true=publish, false=unpublish)
      required:
        - published
    BulkGenerateCertificatesDto:
      type: object
      properties:
        eventId:
          type: string
          description: Event ID to generate certificates for
      required:
        - eventId